// src/resume.schema.ts
export interface ResumeSchema {
// ...
  work?: {
    /**
     * e.g. Facebook
     */
    name?: string;
    /**
     * e.g. Menlo Park, CA
     */
    location?: string;
    /**
     * e.g. Social Media Company
     */
    description?: string;
    /**
     * e.g. Software Engineer
     */
    position?: string;
    /**
     * e.g. http://facebook.example.com
     */
    url?: string;
    /**
     * resume.json uses the ISO 8601 date standard e.g. 2014-06-29
     */
    startDate?: string;
    /**
     * e.g. 2012-06-29
     */
    endDate?: string;
    /**
     * Give an overview of your responsibilities at the company
     */
    summary?: string;
    /**
     * Specify multiple accomplishments
     */
    highlights?: string[];
    [k: string]: any;
  }[];
// ...
  publications?: {
    /**
     * e.g. The World Wide Web
     */
    name?: string;
    /**
     * e.g. IEEE, Computer Magazine
     */
    publisher?: string;
    /**
     * e.g. 1990-08-01
     */
    releaseDate?: string;
    /**
     * e.g. http://www.computer.org.example.com/csdl/mags/co/1996/10/rx069-abs.html
     */
    url?: string;
    /**
     * Short summary of publication. e.g. Discussion of the World Wide Web, HTTP, HTML.
     */
    summary?: string;
    [k: string]: any;
  }[];
  /**
   * List out your professional skill-set
   */
  skills?: {
    /**
     * e.g. Web Development
     */
    name?: string;
    /**
     * e.g. Master
     */
    level?: string;
    /**
     * List some keywords pertaining to this skill
     */
    keywords?: string[];
    [k: string]: any;
  }[];
// ...
}